fix: consolidate session stream files to satisfy pkg file-count gate
Fold session response-stream wiring into runtime_sessions.go and provider
progress publishing into inference_provider.go so pkg/service and
pkg/workers/provider stay within the 15-file package limit.
